+++ svs/migrations/0005_importstructuralvariantbgjob.py
@@ -83,6 +83,6 @@ class Migration(migrations.Migration):
                     ),
                 ),
             ],
-            bases=(variants.models.JobModelMessageMixin2, models.Model),
+            bases=(variants.models.JobModelMessageMixin, models.Model),
         )
     ]