+++ utility/install_chrome.sh
@@ -5,17 +5,17 @@ echo "Installing Chrome + Driver for UI Testing"
 echo "***********************************************"
 
 # Version
-CHROME_DRIVER_VERSION=2.36
+CHROME_DRIVER_VERSION=83.0.4103.39
 
 # Install dependencies
 sudo apt-get update
-sudo apt-get install -y unzip openjdk-8-jre-headless xvfb libxi6 libgconf-2-4
+sudo apt-get install -y unzip xvfb libxi6 libgconf-2-4
 
 # Install Chrome
 wget -N https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb -P ~/
 sudo dpkg -i --force-depends ~/google-chrome-stable_current_amd64.deb
 sudo apt-get -f install -y
-sudo dpkg -i --force-depends ~/google-chrome-stable_current_amd64.deb
+rm ~/google-chrome-stable_current_amd64.deb
 
 # Install ChromeDriver
 wget -N http://chromedriver.storage.googleapis.com/$CHROME_DRIVER_VERSION/chromedriver_linux64.zip -P ~/