+++ variants/migrations/0068_kioskannotatebgjob.py
@@ -68,6 +68,6 @@ class Migration(migrations.Migration):
                     ),
                 ),
             ],
-            bases=(variants.models.JobModelMessageMixin2, models.Model),
+            bases=(variants.models.JobModelMessageMixin, models.Model),
         )
     ]